rename simulator
