+++ Cargo.toml
@@ -50,7 +50,6 @@ members = [
   "misc/load-testing",
   "misc/crowdloan-verifier",
   "xcm-simulator",
-  "example"
 ]
 exclude = ["scripts/check-off-on-deviation"]
 default-members = [
@@ -76,7 +75,6 @@ default-members = [
   "pallets/rewards/rpc/runtime-api",
   "rpc/swap",
   "xcm-simulator",
-  "example"
 ]
 
 [workspace.dependencies]
