+++ nodes/mainnet/src/chain_spec.rs
@@ -74,7 +74,8 @@ pub(crate) fn session_keys(
 }
 
 fn udon_testnet_config_genesis() -> RuntimeGenesisConfig {
-	let seed = "owner word vocal dose decline sunset battle example forget excite gentle waste//";
+	let seed =
+		"owner word vocal dose decline sunset battle xcm-simulator forget excite gentle waste//";
 	let mut initial_authorities: Vec<(
 		AccountId,
 		AccountId,
